Revert "Session auth client (#435)" (#449)
This reverts commit dddc09b70fdb4a46b2b6c5b8cb6d51aad78852a1.
