+++ api/requirements/dev.txt
@@ -10,13 +10,8 @@ charset-normalizer==2.0.12
 coreapi==2.3.3
 coreschema==0.0.4
 coverage==6.3.2
-cryptography==36.0.1
-defusedxml==0.7.1
-dj-rest-auth==2.2.3
 Django==3.2.12
-django-allauth==0.49.0
 django-boto==0.3.12
-django-cors-headers==3.11.0
 django-debug-toolbar==3.2.4
 django-environ==0.8.1
 django-extensions==3.1.5
@@ -39,23 +34,19 @@ MarkupSafe==2.1.0
 mypy==0.931
 mypy-extensions==0.4.3
 numpy==1.22.2
-oauthlib==3.2.0
 packaging==21.3
 pandas==1.4.1
 pluggy==1.0.0
 psycopg2-binary==2.9.3
 py==1.11.0
 pycparser==2.21
-PyJWT==2.3.0
 pynmea2==1.18.0
 pyparsing==3.0.7
 pytest==7.0.1
 pytest-django==4.5.2
 python-dateutil==2.8.2
-python3-openid==3.2.0
 pytz==2021.3
 requests==2.27.1
-requests-oauthlib==1.3.1
 ruamel.yaml==0.17.21
 ruamel.yaml.clib==0.2.6
 six==1.16.0
