+++ api/requirements/prod.txt
@@ -7,13 +7,8 @@ cffi==1.15.0
 charset-normalizer==2.0.12
 coreapi==2.3.3
 coreschema==0.0.4
-cryptography==36.0.1
-defusedxml==0.7.1
-dj-rest-auth==2.2.3
 Django==3.2.12
-django-allauth==0.49.0
 django-boto==0.3.12
-django-cors-headers==3.11.0
 django-environ==0.8.1
 django-extensions==3.1.5
 django-model-utils==4.2.0
@@ -29,19 +24,15 @@ itypes==1.2.0
 Jinja2==3.0.3
 MarkupSafe==2.1.0
 numpy==1.22.2
-oauthlib==3.2.0
 packaging==21.3
 pandas==1.4.1
 psycopg2-binary==2.9.3
 pycparser==2.21
-PyJWT==2.3.0
 pynmea2==1.18.0
 pyparsing==3.0.7
 python-dateutil==2.8.2
-python3-openid==3.2.0
 pytz==2021.3
 requests==2.27.1
-requests-oauthlib==1.3.1
 ruamel.yaml==0.17.21
 ruamel.yaml.clib==0.2.6
 six==1.16.0
