--- api/woeip/apps/core/forms.py
@@ -1,2 +1,0 @@
-from django import forms
-from django.contrib.auth.forms import UserChangeForm, UserCreationForm
