+++ api/woeip/urls.py
@@ -18,14 +18,13 @@ router.register(r"pollutant", views.PollutantViewSet)
 router.register(r"pollutant_values", views.PollutantValueViewSet)
 router.register(r"sensors", views.SensorViewSet)
 router.register(r"timegeo", views.TimeGeoViewSet)
-# router.register(r"users", core_views.UserViewSet) # Disable to prevent personal information leaking
+router.register(r"users", core_views.UserViewSet)
 
 
 urlpatterns = [
     path("", include(router.urls)),
     path("admin/", admin.site.urls),
-    path("auth/", include("dj_rest_auth.urls")),
-    path("auth/register/", include('dj_rest_auth.registration.urls')),
+    path("api-auth/", include("rest_framework.urls", namespace="rest_framework")),
 ]
 
 urlpatterns += swagger_urlpatterns
