+++ nginx-proxy/cors.conf
@@ -1,0 +1,4 @@
+add_header 'Access-Control-Allow-Origin' '*' always;
+add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS';
+add_header 'Access-Control-Allow-Headers' 'User-Agent,Keep-Alive,Content-Type';
+add_header 'Access-Control-Expose-Headers' 'Content-Length,Content-Range';
