+++ web/src/App.test.js
@@ -11,43 +11,6 @@ import { App } from "./App";
 jest.mock("./components/map/box", () => () => <></>);
 
 describe("Application should includes parts of each major section", () => {
-  beforeEach(() => {
-    const fakeLocalStorage = (function () {
-      let store = {};
-
-      return {
-        getItem: function (key) {
-          return store[key] || null;
-        },
-        setItem: function (key, value) {
-          store[key] = value.toString();
-        },
-        removeItem: function (key) {
-          delete store[key];
-        },
-        clear: function () {
-          store = {};
-        },
-      };
-    })();
-
-    Object.defineProperty(window, "localStorage", {
-      value: fakeLocalStorage,
-    });
-  });
-
-  afterEach(() => {
-    localStorage.clear();
-  });
-
-  it("should not show a login screen", () => {
-    localStorage.setItem("authToken", "fakeToken");
-    render(<App />);
-    expect(
-      screen.queryByText(/Sign in to upload data/)
-    ).not.toBeInTheDocument();
-  });
-
   it("should have part of the header", () => {
     render(<App />);
     expect(screen.getByText(/Upload/).href).toMatch("/upload");
@@ -61,10 +24,5 @@ describe("Application should includes parts of each major section", () => {
   it("should have part of the footer", () => {
     render(<App />);
     expect(screen.getByText(/WOEIP/).href).toEqual("https://woeip.org/");
-  });
-
-  it("should show a login screen", () => {
-    render(<App />);
-    expect(screen.getByText(/Sign in to upload data/)).toBeInTheDocument();
   });
 });
