--- web/src/components/auth/login.css
@@ -1,7 +1,0 @@
-.login-container {
-    margin-top: 2rem;
-}
-
-.login-instructions {
-    margin-bottom: 1rem;
-}
