--- web/src/components/auth/register.css
@@ -1,7 +1,0 @@
-.register-container {
-    margin-top: 2rem;
-}
-
-.register-instructions {
-    margin-bottom: 1rem;
-}
