--- web/src/components/auth/tokenContext.js
@@ -1,7 +1,0 @@
-import { createContext } from "react";
-
-export const AuthTokenContext = createContext({
-  authToken: "",
-  tokenLoading: false,
-  setAuthToken: () => {},
-});
