+++ web/src/components/home/index.js
@@ -1,10 +1,9 @@
-import { AuthBoundary } from "../auth/boundary";
-import Welcome from "./welcome";
+import { Container } from "../ui";
 
 export const Home = () => {
   return (
-    <AuthBoundary>
-      <Welcome />
-    </AuthBoundary>
+    <Container textAlign="center">
+      <h2>Welcome to West Oakland Air Quality!</h2>
+    </Container>
   );
 };
