--- web/src/components/home/welcome.css
@@ -1,3 +1,0 @@
-.welcome-intro {
-    max-width: 30rem;
-}
