+++ web/src/components/ui/index.js
@@ -17,7 +17,6 @@ export {
   Icon,
   Label,
   Grid,
-  Form,
 } from "semantic-ui-react";
 
 export const LinkedHeader = ({ ...props }) => (
